Update index.html
Updated talk descriptions using the list of abstracts at the end of the page. I've commented out the list of abstracts - you might want to delete that properly once you've checked I haven't made any errors that stop it compiling. Not all talks have descriptions in the list, and there are some talk abstracts that aren't in the schedule of talks (I've saved a list of them if you want it, but I'm guessing these were talks that were submitted but not given, so probably not of interest).